third_party/WebKit/LayoutTests/inspector/elements/elements-panel-shadow-selection-on-refresh.html

<html>
<head>
<script src="../../http/tests/inspector/inspector-test.js"></script>
<script src="../../http/tests/inspector/elements-test.js"></script>
<script>

function test()
{

    InspectorTest.runTestSuite([
        function setup(next)
        {
            WebInspector.settings.showUAShadowDOM.set(true);
            InspectorTest.expandElementsTree(next);
        },

        function testAuthorShadowRoot(next)
        {
            InspectorTest.findNode(isAuthorShadowRoot, selectReloadAndDump.bind(null, next));
        },

        function testUserAgentShadowRoot(next)
        {
            InspectorTest.findNode(isUserAgentShadowRoot, selectReloadAndDump.bind(null, next));
        },

        function testAuthorShadowRootChild(next)
        {
            InspectorTest.findNode(isAuthorShadowRootChild, selectReloadAndDump.bind(null, next));
        },

        function testUserAgentShadowRootChild(next)
        {
            InspectorTest.findNode(isUserAgentShadowRootChild, selectReloadAndDump.bind(null, next));
        }
    ]);

    function isAuthorShadowRoot(node)
    {
        return node && node.shadowRootType() === WebInspector.DOMNode.ShadowRootTypes.Author;
    }

    function isUserAgentShadowRoot(node)
    {
        return node && node.shadowRootType() === WebInspector.DOMNode.ShadowRootTypes.UserAgent;
    }

    function isAuthorShadowRootChild(node)
    {
        return isAuthorShadowRoot(node.parentNode);
    }

    function isUserAgentShadowRootChild(node)
    {
        return isUserAgentShadowRoot(node.parentNode);
    }

    function selectReloadAndDump(next, node)
    {
        InspectorTest.selectNode(node).then(step0);

        function step0()
        {
            InspectorTest.reloadPage(step1);
        }

        function step1()
        {
            InspectorTest.runAfterPendingDispatches(step2);
        }

        function step2()
        {
            dumpSelectedNode();
            next();
        }

        function dumpSelectedNode()
        {
            var selectedElement = InspectorTest.firstElementsTreeOutline().selectedTreeElement;
            var nodeName = selectedElement ? selectedElement.representedObject.nodeNameInCorrectCase() : "null";
            InspectorTest.addResult("Selected node: '" + nodeName + "'");
        }
    }
}

</script>
</head>

<body onload="runTest()">
<p>
Tests that elements panel preserves selected shadow DOM node on page refresh.
</p>
<span id="hostElement"></span>
<script>
var root = document.getElementById("hostElement").createShadowRoot();
root.innerHTML = "<input type='text'>";
</script>
</body>
</html>
